mv class definition and remove check
