+++ fractal_server/app/schemas/v2/__init__.py
@@ -1,4 +1,3 @@
-from .collection_state import CollectionStateReadV2  # noqa F401
 from .dataset import DatasetCreateV2  # noqa F401
 from .dataset import DatasetExportV2  # noqa F401
 from .dataset import DatasetImportV2  # noqa F401
@@ -23,6 +22,7 @@ from .task import TaskImportV2  # noqa F401
 from .task import TaskLegacyReadV2  # noqa F401
 from .task import TaskReadV2  # noqa F401
 from .task import TaskUpdateV2  # noqa F401
+from .task_collection import CollectionStateReadV2  # noqa F401
 from .task_collection import CollectionStatusV2  # noqa F401
 from .task_collection import TaskCollectPipV2  # noqa F401
 from .workflow import WorkflowCreateV2  # noqa F401
